+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add index alias support ([#50](https://github.com/hasura/ndc-elasticsearch/pull/50))
+
 ## [1.1.3]
 
 ### Changed